feat: Improve handling of command arguments with spaces
This commit introduces the following changes:

1. Add a new function `BuildArgs` in the `pkg/cmdx/args_test.go` file to handle the parsing and processing of command arguments, with a focus on correctly handling arguments that contain spaces.
2. Implement the `GenerateCommand` function in the `pkg/cmdx/command.go` file to generate a `types.DaggerCMD` slice from a command and its arguments, ensuring that arguments with spaces are properly handled.
3. Implement the `ConvertCMDToString` function in the `pkg/cmdx/command.go` file to convert a `types.DaggerCMD` slice back to a string representation of the command, preserving the original formatting and handling of arguments with spaces.

These changes improve the overall command handling functionality in the Daggerx project, making it more robust and flexible when dealing with arguments that contain spaces.
