+++ pkg/cmdx/args_test.go
@@ -1,4 +1,4 @@
-package cmdbuilder
+package cmdx
 
 import (
 	"reflect"
